Enhancements36

+++ install_clouderamanagerserver.sh
@@ -55,10 +55,10 @@ check_root() {
 # Function to discover basic OS details.
 discover_os() {
   if command -v lsb_release >/dev/null; then
-    # CentOS, Ubuntu, RedHatEnterpriseServer, Debian, SUSE LINUX
+    # CentOS, Ubuntu, RedHatEnterpriseServer, RedHatEnterprise, Debian, SUSE LINUX, OracleServer
     # shellcheck disable=SC2034
     OS=$(lsb_release -is)
-    # CentOS= 6.10, 7.2.1511, Ubuntu= 14.04, RHEL= 6.10, 7.5, SLES= 11
+    # CentOS= 6.10, 7.2.1511, Ubuntu= 14.04, RHEL= 6.10, 7.5, SLES= 11, OEL= 7.6
     # shellcheck disable=SC2034
     OSVER=$(lsb_release -rs)
     # 7, 14
@@ -72,29 +72,66 @@ discover_os() {
       if [ -f /etc/centos-release ]; then
         # shellcheck disable=SC2034
         OS=CentOS
-        # 7.5.1804.4.el7.centos, 6.10.el6.centos.12.3
         # shellcheck disable=SC2034
-        OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}.%{RELEASE}\n' | awk -F. '{print $1"."$2}')
+        OSREL=$(rpm -qf /etc/centos-release --qf='%{VERSION}\n' | awk -F. '{print $1}')
         # shellcheck disable=SC2034
-        OSREL=$(rpm -qf /etc/centos-release --qf='%{VERSION}\n')
+        OSNAME=$(awk -F"[()]" '{print $2}' /etc/centos-release | sed 's| ||g')
+        if [ -z "$OSNAME" ]; then
+          # shellcheck disable=SC2034
+          OSNAME="n/a"
+        fi
+        if [ "$OSREL" -le "6" ]; then
+          # 6.10.el6.centos.12.3
+          # shellcheck disable=SC2034
+          OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}.%{RELEASE}\n' | awk -F. '{print $1"."$2}')
+        elif [ "$OSREL" == "7" ]; then
+          # 7.5.1804.4.el7.centos
+          # shellcheck disable=SC2034
+          OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}.%{RELEASE}\n' | awk -F. '{print $1"."$2"."$3}')
+        elif [ "$OSREL" == "8" ]; then
+          if [ "$(rpm -qf /etc/centos-release --qf='%{NAME}\n')" == "centos-stream-release" ]; then
+            # shellcheck disable=SC2034
+            OS=CentOSStream
+            # shellcheck disable=SC2034
+            OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}\n' | awk -F. '{print $1}')
+          else
+            # shellcheck disable=SC2034
+            OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}.%{RELEASE}\n' | awk -F. '{print $1"."$2"."$4}')
+          fi
+        else
+          # shellcheck disable=SC2034
+          OS=CentOSStream
+          # shellcheck disable=SC2034
+          OSVER=$(rpm -qf /etc/centos-release --qf='%{VERSION}\n')
+        fi
+      elif [ -f /etc/oracle-release ]; then
+        # shellcheck disable=SC2034
+        OS=OracleServer
+        # 7.6
+        # shellcheck disable=SC2034
+        OSVER=$(rpm -qf /etc/oracle-release --qf='%{VERSION}\n')
+        # shellcheck disable=SC2034
+        OSNAME="n/a"
       else
         # shellcheck disable=SC2034
         OS=RedHatEnterpriseServer
-        # 7.5, 6Server
+        # 8.6, 7.5, 6Server
         # shellcheck disable=SC2034
         OSVER=$(rpm -qf /etc/redhat-release --qf='%{VERSION}\n')
+        # shellcheck disable=SC2034
+        OSREL=$(echo "$OSVER" | awk -F. '{print $1}')
         if [ "$OSVER" == "6Server" ]; then
           # shellcheck disable=SC2034
           OSVER=$(rpm -qf /etc/redhat-release --qf='%{RELEASE}\n' | awk -F. '{print $1"."$2}')
+        elif [ "$OSREL" == "8" ]; then
           # shellcheck disable=SC2034
-          OSNAME=Santiago
-        else
-          # shellcheck disable=SC2034
-          OSNAME=Maipo
+          OS=RedHatEnterprise
         fi
         # shellcheck disable=SC2034
-        OSREL=$(echo "$OSVER" | awk -F. '{print $1}')
+        OSNAME=$(awk -F"[()]" '{print $2}' /etc/redhat-release | sed 's| ||g')
       fi
+      # shellcheck disable=SC2034
+      OSREL=$(echo "$OSVER" | awk -F. '{print $1}')
     elif [ -f /etc/SuSE-release ]; then
       if grep -q "^SUSE Linux Enterprise Server" /etc/SuSE-release; then
         # shellcheck disable=SC2034
@@ -103,7 +140,7 @@ discover_os() {
       # shellcheck disable=SC2034
       OSVER=$(rpm -qf /etc/SuSE-release --qf='%{VERSION}\n' | awk -F. '{print $1}')
       # shellcheck disable=SC2034
-      OSREL=$(rpm -qf /etc/SuSE-release --qf='%{VERSION}\n' | awk -F. '{print $1}')
+      OSREL=$(echo "$OSVER" | awk -F. '{print $1}')
       # shellcheck disable=SC2034
       OSNAME="n/a"
     fi
@@ -191,7 +228,7 @@ echo "*** $(basename "$0") $*"
 echo "********************************************************************************"
 # Check to see if we are on a supported OS.
 discover_os
-if [ "$OS" != RedHatEnterpriseServer ] && [ "$OS" != CentOS ] && [ "$OS" != Debian ] && [ "$OS" != Ubuntu ]; then
+if [ "$OS" != RedHatEnterpriseServer ] && [ "$OS" != CentOS ] && [ "$OS" != OracleServer ] && [ "$OS" != Debian ] && [ "$OS" != Ubuntu ]; then
   echo "ERROR: Unsupported OS."
   exit 3
 fi
@@ -233,7 +270,7 @@ fi
 echo "Installing Cloudera Manager Server..."
 echo "CM database is: $INSTALLDB"
 echo "CM version is: $SCMVERSION"
-if [ "$OS" == RedHatEnterpriseServer ] || [ "$OS" == CentOS ]; then
+if [ "$OS" == RedHatEnterpriseServer ] || [ "$OS" == CentOS ] || [ "$OS" == OracleServer ]; then
   # Test to see if JDK 6 is present.
   if rpm -q jdk >/dev/null; then
     HAS_JDK=yes
@@ -294,6 +331,10 @@ if [ "$OS" == RedHatEnterpriseServer ] || [ "$OS" == CentOS ]; then
       exit 10
     fi
   fi
+  if [ "$OSREL" -eq 8 ]; then
+    # https://developers.redhat.com/blog/2018/11/14/python-in-rhel-8-3#why_not_just__python__
+    alternatives --set python /usr/bin/python2
+  fi
   if [ "$INSTALLDB" == embedded ]; then
     yum -y -e1 -d1 install cloudera-manager-server-db-2
     service cloudera-scm-server-db start
@@ -305,10 +346,12 @@ if [ "$OS" == RedHatEnterpriseServer ] || [ "$OS" == CentOS ]; then
     chkconfig cloudera-scm-server on
   else
     if [ "$INSTALLDB" == mysql ]; then
-      yum -y -e1 -d1 install mysql-connector-java
-      # Removes JDK 6 if it snuck onto the system. Tests for the actual RPM
-      # named "jdk" to keep virtual packages from causing a JDK 8 uninstall.
-      if [ "$HAS_JDK" == no ] && rpm -q jdk >/dev/null; then yum -y -e1 -d1 remove jdk; fi
+      if [ ! -f /usr/share/java/mysql-connector-java.jar ]; then
+        yum -y -e1 -d1 install mysql-connector-java
+        # Removes JDK 6 if it snuck onto the system. Tests for the actual RPM
+        # named "jdk" to keep virtual packages from causing a JDK 8 uninstall.
+        if [ "$HAS_JDK" == no ] && rpm -q jdk >/dev/null; then yum -y -e1 -d1 remove jdk; fi
+      fi
     elif [ "$INSTALLDB" == postgresql ]; then
       yum -y -e1 -d1 install postgresql-jdbc
     elif [ "$INSTALLDB" == oracle ]; then
